% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgo_points.R
\encoding{UTF-8}
\name{sgo_coordinates}
\alias{sgo_coordinates}
\title{Extracts coordinates from an \code{sgo_points} object}
\usage{
sgo_coordinates(x)
}
\arguments{
\item{x}{An instance of \code{sgo_points}.}
}
\value{
A matrix with 2 or 3 named columns.
}
\description{
Extract the coordinates of an \code{sgo_points} object expressed as a matrix.
}
\examples{
p <- sgo_points(list(57.47777, -4.22472), epsg=4326)
coords <- sgo_coordinates(p)

}
