% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgs_area.R
\encoding{UTF-8}
\name{sgs_area}
\alias{sgs_area}
\title{Calculate areas from a set of points}
\usage{
sgs_distance(x, ...)
}
\arguments{
\item{x}{A \code{sgs_points} object describing a set of points (clockwise, countercockwise?)}
}
\value{
Value of the are in squared metres.
}
\description{
Calculates the planar area for a set of points defined in the OS NGR or
ETRS89-LAEA. The geodetic area is calculated when the entered points are
expressed in angular coordinates.
}
\details{
TODO.
}
\examples{
#TODO
}
