% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgs_wgs84.R
\encoding{UTF-8}
\name{sgs_en_wgs84}
\alias{sgs_en_wgs84}
\title{Pseudo - Mercator to WGS84 Longitude/Latitude}
\usage{
sgs_en_wgs84(x, to = 4326)
}
\arguments{
\item{x}{A \code{sgs_points} object describing a set of points in the
geodetic coordinate system EPSG=3857.}

\item{to}{Numeric. Sets the \code{epsg} code of the destination Geodetic
Coordinate System. 4326 (WGS84) by default. And currently doesn't support any
other.}
}
\value{
An object of class \code{sgs_points} whose coordinates are defined as
Longitude/Latitude.
}
\description{
Converts Pseudo - Mercator coordinates to WGS84 (EPSG=4326)
}
\details{
Currently converts ONLY from EPSG 3857 to 4326 (Longitude/Latitude).
}
\examples{
p <- sgs_points(list(-11169055.58, 2810000.00), epsg=3857)
res <- sgs_en_wgs84(p)
}
\references{
IOGP Publication 373-7-2 - Geomatics Guidance Note number 7,
part 2 (October 2020) \url{https://epsg.org/guidance-notes.html}
}
\seealso{
\code{\link{sgs_points}}, \code{\link{sgs_wgs84_en}}.
}
