% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgs.R
\encoding{UTF-8}
\name{sgs_equal}
\alias{sgs_equal}
\title{TODO}
\usage{
sgs_equal(x, y, tolerance = NULL, core.only=TRUE)
}
\arguments{
\item{x}{A \code{sgs_points} object.}

\item{y}{A \code{sgs_points} object.}

\item{tolerance}{TODO. 2-elemntes vector to set coordinates differencfe tolerance}

\item{core.only}{TODO. if evaluate only core columns of sgs_points}
}
\value{
TODO. TRUE or a vector of mode 'character' describing the differences between x and y.
}
\description{
TODO.
}
\details{
TODO.
}
