% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgs_laea.R
\encoding{UTF-8}
\name{sgs_laea_etrs}
\alias{sgs_laea_etrs}
\title{ETRS89-LAEA Easting/Northing to ETRS89 angular coordinates}
\usage{
sgs_laea_etrs(x)
}
\arguments{
\item{x}{A \code{sgs_points} object describing a set of points in the
projected coordinate system EPSG=3035.}
}
\value{
An object of class \code{sgs_points} whose coordinates are defined as
Longitude/Latitude in the CRS ETRS89.
}
\description{
Converts ETRS89-LAEA Easting/Northing to ETRS89 geodetic coordinates
(EPSG:4258)
}
\details{
ETRS89-LAEA (EPSG:3035) is a CRS for pan-European statistical mapping at all
scales or other purposes where true area representation is required.
}
\examples{
p <- sgs_points(list(3962799.45, 2999718.85), epsg=3035)
res <- sgs_laea_etrs(p)
}
\references{
IOGP Publication 373-7-2 - Geomatics Guidance Note number 7,
part 2 (October 2020) \url{https://epsg.org/guidance-notes.html}
}
\seealso{
\code{\link{sgs_points}}, \code{\link{sgs_area}}.
}
