% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgs_perimeter.R
\encoding{UTF-8}
\name{sgs_perimeter}
\alias{sgs_perimeter}
\title{Calculate area from an ordered set of points}
\usage{
sgs_perimeter(x,
  grid.true.distance = ifelse(isTRUE(x$epsg==27700 || x$epsg==7405),
  TRUE, FALSE))
}
\arguments{
\item{x}{A \code{sgs_points} object describing an ordered set of points. It
also accepts a list of sgs_points objects each of them describing a
different polygon (area).}

\item{grid.true.distance}{TODO. BNG only}
}
\value{
When entering a single \code{sgs_points} object it returns the value of the
area in squared metres. For a list of \code{sgs_points} objects, it will
return a list of areas in squared metres.
}
\description{
Calculates the planar area for a set of points defined in the OS BNG or
ETRS89-LAEA. An approximation of the geodetic area is calculated when
entered points are expressed in angular coordinates.
}
\details{
Calculate areas using the Gauss's area formula
(https://en.wikipedia.org/wiki/Shoelace_formula).

When using angular coordinates the function performs an approximation of the
geodetic area following the methodology discussed in Berk & Ferlan (2018)
where the area on the ellipsoid is determined by using a region-adapted
equal-area projection (Albers Equal-Area Conic) with one standard parallel.
The standard parallel and the projection origin are tied to the moment
centroid of the polygon.

To reduce the error introduced by boundary simplification and provide an
even more accurate area computation, the boundary segments can be divided by
interpolating vertices on the projected geodesic.
}
\examples{
#TODO
}
\references{
Sandi Berk & Miran Ferlan, 2018. \emph{Accurate area determination in the
cadaster: case study of Slovenia}. Cartography and Geographic Information
Science, 45:1, 1-17. DOI: 10.1080/15230406.2016.1217789

Snyder, J.P. 1987. \emph{Map Projections — A Working Manual}. US Geological
Survey Professional Paper, no. 1395. Washington, DC: US Government Printing
Office. DOI: 10.3133/pp1395
}
