% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgs_points.R
\encoding{UTF-8}
\name{sgs_points}
\alias{sgs_points}
\title{Object containing 2D point coordinates}
\usage{
sgs_points(x, coords=NULL, epsg = NULL)
}
\arguments{
\item{x}{A list or dataframe with at least 2 columns of either
easting/northing or longitude/latitude coordinates per row - depending on
\code{epsg}. \strong{Please note} that the order is important: lat/lon or
northing/easting will produce wrong results.

It also accepts objects of class \code{sf POINT/MULTIPOINT} or
\code{sfc POINT/MULTIPOINT}.}

\item{coords}{A vector with the names of the columns containing the
easting/northing (or longitude/latitude) coordinates.}

\item{epsg}{Specifies the EPSG code of coordinates to store. It can take any
of the following values: \code{4326}, \code{3857}, \code{4277}, \code{27700}
or \code{4258}.}
}
\value{
An object of class \code{sgs_points}.
}
\description{
2D coordinates (and other attributes) of a point or collection of points
}
\details{
This object stores 2D point coordinates and any other column-list of
attributes related to each point. Note that additional column-lists will be
expanded with \code{NA} values if they contain less elements than coordinates.
Currently it only supports 5 \code{epsg}s for coordinates:
\itemize{
\item\code{4326}: WGS84, geodetic coordinate system. The 2 columns in
\code{x} be defined as Longitude and Latitude values respectively. The
defined datum for this set of coordinates is WGS84 (https://epsg.io/4326)

\item\code{3857}: WGS 84 / Pseudo-Mercator, projected coordinate system. The
2 columns in \code{x} must be defined as Easting and Northing values
respectively. The defined datum for this set of coordinates is WGS84
(https://epsg.io/3857)

\item\code{4277}: OSGB36, geodetic coordinate system. The 2 columns in
\code{x} must be defined as Longitude and Latitude values respectively. The
defined datum for this set of coordinates is OSGB 1936
(https://epsg.io/4277). \emph{Coordinates defined like this should only be
used for historical reasons.}

\item\code{27700}: British National Grid, projected coordinate system. The 2
columns in \code{x} must be defined as Easting and Northing values
respectively. The defined datum for this set of coordinates is OSGB 1936
(https://epsg.io/27700).

\item\code{4258}: ETRS89, geodetic coordinate system. The 2 columns in
\code{x} must be defined as Longitude and Latitude values respectively. The
defined datum for this set of coordinates is ETRS89 (https://epsg.io/4258).
}
}
\examples{
# lists:
p1 <- sgs_points(list(-3.9369, 56.1165), epsg=4326)
lon <- c(-4.25181,-3.18827)
lat <- c(55.86424, 55.95325)
p2 <- sgs_points(list(longitude=lon, latitude=lat), epsg=4326)
#p3 will expand the list 'desc' up to 2 elements with NA's
# dataframe:
ln <- c(-4.22472, -2.09908)
lt <- c(57.47777, 57.14965)
n <- c("Inverness", "Aberdeen")
df <- data.frame(n, ln, lt, stringsAsFactors = FALSE)
p4 <- sgs_points(df, coords=c("ln", "lt"), epsg=4326)

\dontrun{
library(maps)
map('world', regions=('uk'))
points(x=p1$longitude, y=p1$latitude, pch=0, col="green")
points(x=p2$lon, y=p2$lat, pch=0, col="blue")
points(x=p4$lon, y=p4$lat, pch=0, col="red")
}
}
\seealso{
\code{\link{sgs_polygons}}, \code{\link{sgs_points_xy}},
\code{\link{sgs_points_sf}}, \code{\link{sgs_set_gcs}}.
}
