% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\encoding{UTF-8}
\name{sgs_points_sf}
\alias{sgs_points_sf}
\title{Convert sgs_points objects to sf objects}
\usage{
sgs_points_sf(x)
}
\arguments{
\item{x}{An instance of \code{sgs_points}.}
}
\value{
An object of class \code{sf}.
}
\description{
Converts \code{sgs_points} objects to instances of class \code{sf}. \pkg{sf}
is a package that allows working with spatial vector data.

The reverse conversion (from \code{sf} to \code{sgs_points}) is done
automatically by the \code{\link{sgs_points}} contructor.
}
\examples{
p <- sgs_points(list(57.47777, -4.22472), epsg=4326)
sf.p <- sgs_points_sf(p)
sf::st_crs(sf.p)
}
