% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgs_points.R
\encoding{UTF-8}
\name{sgs_points_xy-deprecated}
\alias{sgs_points_xy-deprecated}
\title{Extracts coordinates from an \code{sgs_points} object}
\usage{
sgs_points_xy(x)
}
\arguments{
\item{x}{An instance of \code{sgs_points}.}
}
\value{
A matrix with 2 named columns.
}
\description{
Extract the coordinates of an \code{sgs_points} object expressed as a matrix.
}
\examples{
\dontrun{
p <- sgs_points(list(57.47777, -4.22472), epsg=4326)
coords <- sgs_points_xy(p)
}
}
\seealso{
\code{\link{sgs-deprecated}}
}
\keyword{internal}
