% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgs_polygons.R
\encoding{UTF-8}
\name{sgs_polygons}
\alias{sgs_polygons}
\title{Create sf object from list of coordinates or wkb}
\usage{
sgs_polygons(x, epsg = NULL, wkb = NULL, coords = NULL, by = NULL, buffer = NULL)
}
\arguments{
\item{x}{A dataframe.}

\item{epsg}{Numeric. Sets the \code{epsg} code of the coordinates (or WKB field)
present int he dataframe.}

\item{wkb}{Character. Name of the column containing WKB data. If this parameter is
informed then the parameter \code{coords} is ignored.}

\item{coords}{Vector of chracters containing the names of the columns with
coordinates information. \code{wkb} must be NULL when using this paramater.}

\item{by}{Character. Name of the column that will be used to group series of points
into a single polygon. Ignored when using the paramater \code{wkb}.}

\item{buffer}{Numeric or character. If numeric, then it defines the buffer to apply to each polygon.
If it is a string then it must contain the name of the column having buffer distances
to apply to each polygon. In this case every row of the column must contain a number because
when grouping points with \code{by} the routine will also aggregate buffer distances by the
same column.
\code{buffer} must be expressed in the same units as the coordinates (decimal degrees or meters).
Take into account that \code{sf} does not correctly buffer longitude/latitude data.}
}
\value{
An object of class \code{sf} containing the corresponding geometries
(and IDs when there are multiple polygons).
}
\description{
Creates an sf object containing a collection of polygons (planar surfaces defined
by one exterior boundary and zero or more interior boundaries) from a list of
coordinates or WKB fields. This object facilitates all the operations with polygons
in this package.
}
\examples{
df <- data.frame(latitude=c(55.86424,55.95325,55.87426,56.87426,55.75426),
                 longitude=c(-4.25181,-3.18827,-3.78690,-2.78690,-3.54690),
                 id = c(1,1,1,2,2))
# create a single polygon disregarding different id's
polygon <- sgs_polygons(df, epsg=4326, coords=c("latitude","longitude"))

# create a set of polygons, identified by a column in df
polygons <- sgs_polygons(df, epsg=4326, coords=c("latitude","longitude"), by="id")
}
\seealso{
\code{\link{sgs_points}}.
}
