% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgs_transform.R
\encoding{UTF-8}
\name{sgs_transform}
\alias{sgs_transform}
\title{Coordinate transformation of a set of points}
\usage{
sgs_transform(x, to = NULL, ...)
}
\arguments{
\item{x}{A \code{sgs_points} object.}

\item{to}{Specifies the EPSG code to convert the coordinates to. Currently it
can take any of the following values: \code{4326}, \code{3857}, \code{4277},
\code{27700} or \code{4258}.}

\item{...}{Additional parameters passed to internal functions. #TODO: improve
this description and which parameters may admit...}
}
\value{
An object of class 'sgs_points'.
}
\description{
Transforms the coordinate system of a set of points to any supported
coordinate system.
}
\details{
This function is a wrapper of specific transformation functions
(\code{\link{sgs_bng_lonlat}}, \code{\link{sgs_en_wgs84}},
\code{\link{sgs_lonlat_bng}}, \code{\link{sgs_wgs84_en}},
\code{\link{sgs_lonlat_bng}}, \code{\link{sgs_lonlat_bng}}) that transforms
the coordinate system of a set of points to any of the supported coordinate
systems.
#TODO warning about wgs84/etrs89 transformation
}
\examples{
ln <- c(-4.22472, -2.09908)
lt <- c(57.47777, 57.14965)
n <- c("Inverness", "Aberdeen")
df <- data.frame(n, ln, lt, stringsAsFactors = FALSE)
locations <- sgs_points(df, coords=c("ln", "lt"), epsg=4326)

locations.bng <- sgs_transform(locations, to=27700)
locations.osgb36 <- sgs_transform(locations, to=4277)
locations.ngr <- sgs_bng_ngr(sgs_transform(locations, to=27700))
locations.wgs84EN <- sgs_transform(locations.bng, to=3857)
}
\seealso{
\code{\link{sgs_points}}, \code{\link{sgs_coordinates}},
\code{\link{sgs_set_gcs}}, \code{\link{sgs_bng_ngr}}
}
