% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgs_area.R
\encoding{UTF-8}
\name{sgs_area}
\alias{sgs_area}
\title{Calculate area from an ordered set of points}
\usage{
sgs_area(x, ...)
}
\arguments{
\item{x}{A \code{sgs_points} object describing a set of points (clockwise, countercockwise?)}

\item{...}{Currently ignored}
}
\value{
Value of the are in squared metres.
}
\description{
Calculates the planar area for a set of points defined in the OS NGR or
ETRS89-LAEA. The geodetic area is calculated when entered points are
expressed in angular coordinates.
}
\details{
Calculate areas using the Gauss's area formula (https://en.wikipedia.org/wiki/Shoelace_formula)
}
\examples{
#TODO
}
\references{
Sandi Berk & Miran Ferlan, 2018. \emph{Accurate area determination in the
cadaster: case study of Slovenia}. Cartography and Geographic Information
Science, 45:1, 1-17. DOI: 10.1080/15230406.2016.1217789

Snyder, J.P. 1987. \emph{Map Projections — A Working Manual}. US Geological
Survey Professional Paper, no. 1395. Washington, DC: US Government Printing
Office. DOI: 10.3133/pp1395
}
