% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgs_distance.R
\encoding{UTF-8}
\name{sgs_distance}
\alias{sgs_distance}
\title{Calculate distance(s) between points}
\usage{
sgs_distance(x, y, by.element = FALSE,
  which = ifelse(isTRUE(x$epsg==27700 || x$epsg==7405), "BNG", "Harvesine"),
  grid.distance = ifelse(isTRUE(x$epsg==27700 || x$epsg==7405), TRUE, FALSE),
  iterations = 20L)
}
\arguments{
\item{x}{A \code{sgs_points} object describing a set of points in a geodetic
coordinate system.}

\item{y}{A \code{sgs_points} object, defaults to \code{x}.}

\item{by.element}{Logical variable. If \code{TRUE}, return a vector with
distance between the first elements of \code{x} and \code{y}, the second,
etc. If \code{FALSE}, return the dense matrix with all pairwise distances.}

\item{which}{Character vector. For geodetic coordinates one of
\code{Harvesine} or \code{Vicenty}. For points in OS British National Grid
coordinates it defaults to \code{BNG}.}

\item{grid.distance}{Logical variable. Currently only used for BNG
coordinates. If TRUE it returns the true (geodesic) distance.}

\item{iterations}{Numeric variable. Maximum number of iterations used in the
Vicenty method.}
}
\value{
If \code{by.element} is \code{FALSE} \code{sgs_distance} returns a dense
numeric matrix of dimension length(x) by length(y). Otherwise it returns a
numeric vector of length \code{x} or \code{y}, the shorter one being
recycled. Distances involving empty geometries are \code{NA}.
}
\description{
Calculates the distance between OS National Grid Reference points or using
the Harvesine or Vicenty formulae for points with angular coordinates.
}
\details{
TODO. fro grid.distance explain how corrrection factors work
}
\examples{
#TODO
}
