% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osgrid.R
\encoding{UTF-8}
\name{sgs_ngr_bng}
\alias{sgs_ngr_bng}
\title{NGR to BNG Easting/Northing}
\usage{
sgs_ngr_bng(x, col = NULL)
}
\arguments{
\item{x}{A list or vector containing strings describing OS National Grid
References, with or without whitespace separators. (eg 'SU 387 148').}

\item{col}{Character string with the name of the 'column' containing the
vector of NGR values, it is required when \code{x} is a list with more than
one column.}
}
\value{
An object of class \code{sgs_points} whose coordinates are defined as
Easting/Northing.
}
\description{
Converts OS National Grid References to Easting/Northing coordinates
}
\details{
All entered standard grid references can range from two-digit references up
to 10-digit references (1m × 1m square).
If \code{x} is a list with 2 or more vector elements, \code{col} can be used
to inform the function which of the elements contains the NGR strings. The rest
of the elements will be appended to the resulting object. See examples.
}
\examples{
vec <- c("NN 166 712", "HU 3863 7653")
lst <- list(p1="NN166712", p2="HU 38637653")
v <- sgs_ngr_bng(vec)
l <- sgs_ngr_bng(lst)

# any additional column (here 'attr') will be added to the result
extra <- list(p=c("NN 166712", "HU38637653"),
              attr=c("name1","name2"))
res <- sgs_ngr_bng(extra, col="p")
res

# grid references returned by sgs_bng_ngr are within an
# element (column) named 'ngr'
grid <- sgs_bng_ngr(sgs_points(list(x=247455, y=706338,
                                    name="Ben Venue"),
                                epsg=27700))
bng <- sgs_ngr_bng(grid, col="ngr")
}
\seealso{
\code{\link{sgs_points}}, \code{\link{sgs_bng_ngr}}.
}
