% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgs_transform.R
\encoding{UTF-8}
\name{sgs_transform}
\alias{sgs_transform}
\title{Coordinate transformation of a set of points}
\usage{
sgs_transform(x, to = NULL, ...)
}
\arguments{
\item{x}{A \code{sgs_points} object.}

\item{to}{Specifies the EPSG code to convert the coordinates to. See
\code{\link{sgs_points}} for a list of supported EPSG codes.}

\item{...}{Additional parameters passed to internal functions. Currently it
supports the additional arguments seen in \code{sgs_bng_lonlat} and
\code{sgs_lonlat_bng}.
this description and which parameters may admit...}
}
\value{
An object of class 'sgs_points'.
}
\description{
Transforms the coordinate system of a set of points to any supported
coordinate system.
}
\details{
This function is a wrapper of specific transformation functions
(\code{\link{sgs_bng_lonlat}}, \code{\link{sgs_en_wgs84}},
\code{\link{sgs_lonlat_bng}}, \code{\link{sgs_wgs84_en}},
\code{\link{sgs_laea_etrs}}, \code{\link{sgs_etrs_laea}},
\code{\link{sgs_cart_lonlat}}, \code{\link{sgs_lonlat_cart}}) that transforms
the coordinate system of a set of points to any of the supported coordinate
systems by this package.

Please note that this package assumes that the Coordinate Reference Systems
(CRS) ETRS89 and WGS84 are the same within the UK, but this shouldn't be a
problem for most civilian use of GPS satellites. If a high-precision
transformation between WGS84 and ETRS89 is required then it is recommended
to use a different package to do the conversion.
}
\examples{
ln <- c(-4.22472, -2.09908)
lt <- c(57.47777, 57.14965)
n <- c("Inverness", "Aberdeen")
df <- data.frame(n, ln, lt, stringsAsFactors = FALSE)
locations <- sgs_points(df, coords=c("ln", "lt"), epsg=4326)

locations.bng <- sgs_transform(locations, to=27700)
locations.osgb36 <- sgs_transform(locations, to=4277)
locations.ngr <- sgs_bng_ngr(sgs_transform(locations, to=27700))
locations.wgs84EN <- sgs_transform(locations.bng, to=3857)
}
\seealso{
\code{\link{sgs_points}}, \code{\link{sgs_coordinates}},
\code{\link{sgs_set_gcs}}, \code{\link{sgs_bng_ngr}}
}
